관심있는상품등록 구현완료

+++ src/main/java/com/usedmarket/service/MemberDeleteService.java
@@ -1,10 +1,9 @@
 package com.usedmarket.service;
 
 import com.usedmarket.constant.ItemStatus;
-import com.usedmarket.entity.Item;
-import com.usedmarket.entity.ItemImage;
-import com.usedmarket.repository.ItemImageRepository;
-import com.usedmarket.repository.ItemRepository;
+import com.usedmarket.entity.*;
+import com.usedmarket.repository.*;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,17 +17,22 @@ public class MemberDeleteService {
     private final ItemRepository itemRepository;
     private final ItemImageRepository itemImageRepository;
     private final ItemImageService itemImageService;
+    private final WishlistItemRepository wishlistItemRepository;
+    private final WishlistRepository wishlistRepository;
+    private final WishlistService wishlistService;
 
     public void soldOutStateDelete(Long memberId){
         List<Item> items = itemRepository.findByMemberId(memberId);
         Long itemId;
 
-        for (Item item : items) {
-            if (item.getItemStatus().equals(ItemStatus.SOLD_OUT)) {
-                itemId = item.getId();
-                soldOutImageDelete(itemId);
-                itemRepository.deleteById(item.getId());
+        if(items != null) {
+            for (Item item : items) {
+                if (item.getItemStatus().equals(ItemStatus.SOLD_OUT)) {
+                    itemId = item.getId();
+                    soldOutImageDelete(itemId);
+                    itemRepository.deleteById(item.getId());
 
+                }
             }
         }
     }
@@ -36,8 +40,23 @@ public class MemberDeleteService {
     public void soldOutImageDelete(Long itemId){
         List<ItemImage> itemImages = itemImageRepository.findByItemIdOrderByIdAsc(itemId);
         for (ItemImage itemImage : itemImages) {
-            Long imageId = itemImage.getId();
-            itemImageService.deleteItemImage(imageId);
+            itemImageService.deleteItemImage(itemImage.getId());
+        }
+    }
+
+    public void wishlistDelete(Long memberId){
+        Wishlist wishlist = wishlistRepository.findByMemberId(memberId);
+
+        if(wishlist!=null) {
+            List<WishlistItem> wishlistItems = wishlistItemRepository.findByWishlistId(wishlist.getId());
+
+            if (wishlistItems != null) {
+                for (WishlistItem wishlistItem : wishlistItems) {
+                    wishlistService.deleteWishlist(wishlistItem.getId());
+                }
+            }
+
+            wishlistRepository.deleteById(wishlist.getId());
         }
     }
